Externalize Strings
- Externalize all strings for go-whisk
